refactor: move AuthModal and useAuthModal out of wallet lib
