+++ apps/evm/src/containers/AuthModal/AccountDetails/index.tsx
