+++ apps/evm/src/containers/AuthModal/AccountDetails/styles.ts
