+++ apps/evm/src/containers/AuthModal/index.tsx
@@ -1,7 +1,7 @@
 import { Modal } from 'components/Modal';
+import { useAuthModal } from 'hooks/useAuthModal';
 import { useTranslation } from 'libs/translations';
 import { useAccountAddress } from 'libs/wallet/hooks/useAccountAddress';
-import { useAuthModal } from 'libs/wallet/hooks/useAuthModal';
 import { useChainId } from 'libs/wallet/hooks/useChainId';
 import { useLogIn } from 'libs/wallet/hooks/useLogIn';
 import { useLogOut } from 'libs/wallet/hooks/useLogOut';
