+++ apps/evm/src/hooks/useAuthModal/store/__tests__/index.spec.ts
