+++ apps/evm/src/hooks/useAuthModal/store/index.ts
