+++ apps/evm/src/libs/wallet/Web3Wrapper/index.tsx
@@ -1,7 +1,6 @@
 import { WagmiConfig } from 'wagmi';
 
 import { AuthHandler } from './AuthHandler';
-import { AuthModal } from './AuthModal';
 import config from './config';
 
 export interface Web3WrapperProps {
@@ -10,7 +9,6 @@ export interface Web3WrapperProps {
 
 export const Web3Wrapper: React.FC<Web3WrapperProps> = ({ children }) => (
   <WagmiConfig config={config}>
-    <AuthModal />
     <AuthHandler />
 
     {children}
