+++ apps/evm/src/libs/wallet/index.ts
@@ -13,4 +13,3 @@ export * from './hooks/useSwitchChain';
 export * from './hooks/useChainId';
 export * from './hooks/useLogIn';
 export * from './hooks/useLogOut';
-export * from './hooks/useAuthModal';
